Validacion de datos con api response

+++ app/Http/Controllers/Api/V1/DeveloperController.php
@@ -4,9 +4,10 @@ namespace App\Http\Controllers\Api\V1;
 
 use App\Http\Controllers\Controller;
 use App\Models\Developer;
-use App\Http\Requests\DeveloperRequest;
+//use App\Http\Requests\DeveloperRequest;
 use App\Http\Resources\V1\DeveloperCollection;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class DeveloperController extends Controller
 {
@@ -33,12 +34,37 @@ class DeveloperController extends Controller
      * @param  \Illuminate\Http\DeveloperRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(DeveloperRequest $request)
+    public function store(Request $request)
     {
-        
-        $developer = $this->developer->create($request->all());
+        $data = $request->all();
 
-        return response()->json($developer, 201);
+        $validator = Validator::make($data,[
+            'name'=>'required|regex:/^[\pL\s\-]+$/u|max:255',
+            'profession'=>'required|regex:/^[\pL\s\-]+$/u|max:255',
+            'position'=>'required|regex:/^[\pL\s\-]+$/u|max:255',
+            'technology'=>'required|regex:/^[\pL\s\-]+$/u|max:255'
+             ]);
+        
+             if ($validator->fails()){
+                 return response(
+                     [
+                     'success' => false,
+                     'error' => $validator->errors(),
+                     ], 
+                        400);
+             }
+
+             $developer = Developer::create($data);
+
+             return response([
+                'success' => true,
+                'message' => 'Desarrollador creado correctamente',
+                'desarrollador' => $developer,
+             ], 201);
+
+            
+
+                //return response()->json($developer, 201);
     }
 
     /**
@@ -47,8 +73,16 @@ class DeveloperController extends Controller
      * @param  \App\Models\Developer  $developer
      * @return \Illuminate\Http\Response
      */
-    public function show(Developer $developer)
+    public function show($id)
     {
+       $developer = Developer::where('id',$id)->first();
+
+        if (!$developer) {
+            return response([
+                'success' => false,
+                'message' => 'No se encontro Desarrollador'
+            ], 200);
+        }
         return response()->json($developer);
     }
 
@@ -58,13 +92,66 @@ class DeveloperController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @param  \App\Models\Developer  $developer
      * @return \Illuminate\Http\Response
+     *  
      */
-    public function update(DeveloperRequest $request, Developer $developer)
+
+   /* public function update(DeveloperRequest $request, Developer $developer)
     {
         $developer->update($request->all());
 
         return response()->json($developer);
+    }   */
+    
+   public function update(Request $request, Developer $developer)
+
+   
+    {
+     //  $developer= Developer::find($developer);
+
+      if (!$developer) {
+        return response([
+            'success' => false,
+            'message' => 'No se encontro Desarrollador'
+        ], 200);
     }
+      
+
+      $data = $request->all();
+       
+       $validator = Validator::make($data,[
+           'name'=>'required|regex:/^[\pL\s\-]+$/u|max:255',
+           'profession'=>'required|regex:/^[\pL\s\-]+$/u|max:255',
+            'position'=>'required|regex:/^[\pL\s\-]+$/u|max:255',
+            'technology'=>'required|regex:/^[\pL\s\-]+$/u|max:255'
+            ]);
+
+           if ($validator->fails()){
+                return response(
+                    [
+                    'success' => false,
+                    'error' => $validator->errors(),
+                    ], 
+                       400);
+            } 
+            //$developer->update($data);
+
+
+            //$developer = Developer::update($data);
+
+
+
+            // $developer = Developer::where('id',$id)->update($data);
+
+            $developer->update($data);
+
+            return response([
+                'success' => true,
+                'message' => 'Desarrollador actualizado correctamente',
+                'desarrollador' => $developer,
+             ], 201);
+
+            
+    } 
 
     /**
      * Remove the specified resource from storage.
@@ -72,10 +159,25 @@ class DeveloperController extends Controller
      * @param  \App\Models\Developer  $developer
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Developer $developer)
+    public function destroy($id)
     {
+        $developer= Developer::find($id);
+        
+        if (!$developer) {
+            return response([
+                'success' => false,
+                'message' => 'No se encontro Desarrollador'
+            ], 200);
+        }
+
         $developer->delete();
 
-        return response()->json(null, 204);
+        return response([
+            'success' => true,
+            'message' => 'Desarrollador eliminado',
+            'desarrollador' => $developer,
+         ], 201);
+
+       //return response()->json(null, 204);
     }
 }
